refactor(self-custodial): move the account-mode hook into the self-custodial namespace
