+++ app/self-custodial/hooks/use-self-custodial-account-mode.ts
